updated budgets

+++ flask_app/controllers/budgets.py
@@ -5,6 +5,9 @@ from flask_app.models.budget_item import Budget_Items
 
 @app.route('/budget', methods=['GET'])
 def budget_page():
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     budget = Budget.get_budget()
     print("budget data", budget)
     items = Budget_Items.get_all_items()
@@ -12,6 +15,9 @@ def budget_page():
 
 @app.route('/add-budget', methods=['POST'])
 def add_budget():
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     data = {
         "total": request.form['total'],
         "user_id": 1
@@ -21,6 +27,9 @@ def add_budget():
 
 @app.route('/update-budget', methods=['POST'])
 def update_budget():
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     data = {
         "total": request.form['total'],
         "id": 2  # Assuming you have only one budget entry, or you need to pass the correct budget id
@@ -30,6 +39,9 @@ def update_budget():
 
 @app.route('/add-item', methods=['POST'])
 def add_item():
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     data = {
         "item_name": request.form['item_name'],
         "price": request.form['price'],
@@ -41,6 +53,9 @@ def add_item():
 
 @app.route('/update-item/<int:id>', methods=['POST'])
 def update_item(id):
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     data = {
         "id": id,
         "price": request.form['price']
@@ -51,6 +66,9 @@ def update_item(id):
 
 @app.route('/delete-item/<int:id>')
 def delete_item(id):
+    if "user_id" not in session:
+        flash("Please log in.", "login")
+        return redirect("/")
     data = {"id": id}
     Budget_Items.delete_item(data)
     return redirect('/budget')
